# specs/003-backend-foundations/contracts/api.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot Backend API
  description: API for the RAG chatbot backend with OpenAI Agents SDK integration
  version: 1.0.0
servers:
  - url: https://api.rag-chatbot.com
    description: Production server
  - url: http://localhost:8000
    description: Development server
paths:
...
  /sources/{chunk_id}:
    get:
      summary: Get source information for a chunk
      description: Retrieve the source information for a specific content chunk
      parameters:
        - name: chunk_id
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the content chunk
      responses:
        '200':
          description: Source information for the chunk
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceReference'
        '404':
          description: Chunk not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    SourceReference:
      type: object
      required:
        - chunk_id
        - module
        - chapter
        - url
      properties:
        chunk_id:
          type: string
          description: Reference to the content chunk
          example: "chunk_abc123"
        module:
          type: string
          description: Module where the content appears
          example: "Foundations"
        chapter:
          type: string
          description: Chapter name or identifier
          example: "Introduction to RAG"
        anchor:
          type: string
          description: Specific anchor/section within the source
          example: "section-1.2"
        url:
          type: string
          description: Full URL to the source
          example: "https://book.example.com/chapter-1#section-1.2"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "INVALID_QUERY"
        message:
          type: string
          description: Human-readable error message
          example: "Query must be a non-empty string"
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication